Add uploader, eslint, event modelling
This is a first stab at completing the loop - a view receives a
reference to an event bus where it can publish DOM events. The view
returns a function from a model to a virtual DOM, enabling an event
stream publishing model changes to go to a stream of VDOM patches, and
finally to DOM updates.

I'm still not satisfied with the slisding window, but I'd like to commit
this as it is a reasonable working architecture.